feat: session auth test

+++ backend/my-adonisjs-app/app/controllers/schedules_controller.ts
@@ -5,8 +5,12 @@ export default class SchedulesController {
   /**
    * Display a list of user schedules
    */
-  async index({ params }: HttpContext) {
+  async index({ params, auth }: HttpContext) {
     const userId = params.user_id
+    const authUserId = auth.user?.id
+    if (userId !== authUserId) {
+      return 'Unauthorized access'
+    }
     if (userId) {
       return Schedule.query().where('userId', userId)
     }
@@ -16,7 +20,12 @@ export default class SchedulesController {
   /**
    * Handle form submission for the create action
    */
-  async store({ request }: HttpContext) {
+  async store({ request, auth, params }: HttpContext) {
+    const userId = params.user_id
+    const authUserId = auth.user?.id
+    if (userId !== authUserId) {
+      return 'Unauthorized access'
+    }
     const payload = await request.validateUsing(createScheduleValidator)
     const schedule = await Schedule.create(payload)
     return { message: 'Schedule created.', schedule }
@@ -25,10 +34,15 @@ export default class SchedulesController {
   /**
    * Show schedule with matching courses
    */
-  async show({ params }: HttpContext) {
+  async show({ params, auth }: HttpContext) {
+    const userId = params.user_id
+    const authUserId = auth.user?.id
+    if (userId !== authUserId) {
+      return
+    }
     const schedule = await Schedule.query()
       .where('id', params.id)
-      .andWhere('userId', params.user_id)
+      .andWhere('userId', userId)
       .preload('courses')
       .firstOrFail()
 
@@ -38,7 +52,12 @@ export default class SchedulesController {
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request }: HttpContext) {
+  async update({ params, request, auth }: HttpContext) {
+    const userId = params.user_id
+    const authUserId = auth.user?.id
+    if (userId !== authUserId) {
+      return 'Unauthorized access'
+    }
     const payload = await request.validateUsing(updateScheduleValidator)
     const currSchedule = await Schedule.findOrFail(params.id)
     currSchedule.merge(payload)
@@ -49,7 +68,15 @@ export default class SchedulesController {
    * Delete record
    */
   async destroy(ctx: HttpContext) {
+    const userId = ctx.params.user_id
+    const authUserId = ctx.auth.user?.id
+    if (userId !== authUserId) {
+      return 'Unauthorized access'
+    }
     const schedule = await this.show(ctx)
+    if (!schedule) {
+      return 'Schedule not found'
+    }
     await schedule.delete()
     return { message: 'Schedule successfully deleted.' }
   }
